move the endpoint definitions from main into a new routers module
