+++ app/core/__init__.py
@@ -1,0 +1,1 @@
+'''A file to indicate that the core is a module and can be imported'''
